+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 - Added `Log` variant to `event::Scenario`. ([#258])
 - Added `embeddings` field to `writer::json::Step` and `writer::json::HookResult`. ([#261])
 - Added `report_time` field to `writer::libtest::Cli`. ([#264], [#265])
+- Bumped up [MSRV] to 1.70 for using the `IsTerminal` trait from `std`. ([#288])
 
 ### Added
 
@@ -33,6 +34,7 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 [#265]: /../../pull/265
 [#272]: /../../discussions/272
 [#273]: /../../pull/273
+[#288]: /../../pull/288