+++ codegen/CHANGELOG.md
@@ -11,11 +11,16 @@ All user visible changes to `cucumber-codegen` crate will be documented in this 
 
 [Milestone](/../../milestone/24)
 
+### BC Breaks
+
+- Bumped up [MSRV] to 1.70 for using the `IsTerminal` trait from `std`. ([#288])
+
 ### Changed
 
 - Switched to 2.0 version of [`syn`]. ([#266])
 
 [#266]: /../../pull/266
+[#288]: /../../pull/288